backend/requirements.txt
django==5.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.0.1
channels==4.1.0
